+++ src/data_providers/mod.rs
@@ -1,0 +1,1 @@
+pub mod binance;
